Upgrade to Mermaid 10.6.1
Bump the version of Mermaid up to 10.6.1. Per the release notes at
https://github.com/mermaid-js/mermaid/blob/develop/CHANGELOG.md the
render function is async now, so update to the new async render
function.

There's also apparently a change to use ESM, but I didn't seem to have
any trouble with just bumping the version and updating it to use async.
It looks like it's working without any change to how we load it, but I'm
not much of a frontend developer so I could be missing something.