Update homepage and documentation headings, remove 'Get Insiders' button
Replace the website's home and documentation pages' heading text to reflect a more personal blog versus 'technical docs'. Also resulted in the removal of the 'Get Insiders' button which was not required in the new context. The site description has also been updated to be more relevant to the new blog focus.